+++ frontend/src/components/Auth/SignInForm/SignInForm.tsx
@@ -5,34 +5,29 @@ import { useRouter } from "next/navigation";
 import React from "react";
 import { toast } from "react-toastify";
 
-import { signUp } from "@/api/http/auth/auth";
-import { AuthForm } from "@/components/AuthForm/AuthForm";
-import {
-  fullNameInput,
-  messages,
-  passwordInput,
-  signUpForm,
-  signUpValidationSchema,
-  usernameInput,
-} from "@/components/AuthForms/authForms.config";
-import { axiosErrorHandler, handleUsernameBlur } from "@/components/AuthForms/authForms.utils";
+import { signIn } from "@/api/http/auth/auth";
+import { passwordInput, signInForm, signInValidationSchema, usernameInput } from "@/components/Auth/auth.config";
+import { axiosErrorHandler, handleUsernameBlur } from "@/components/Auth/auth.utils";
 import { CustomInput } from "@/components/CustomInput/CustomInput";
+import { Form } from "@/components/Form/Form";
+import { setAccessToken } from "@/lib/features/auth/authSlice";
+import { useAppDispatch } from "@/lib/hooks";
 
-export function SignUpForm() {
+export function SignInForm() {
   const router = useRouter();
+  const dispatch = useAppDispatch();
   const formik = useFormik({
     initialValues: {
       username: "",
-      fullName: "",
       password: "",
     },
-    validationSchema: signUpValidationSchema,
+    validationSchema: signInValidationSchema,
     onSubmit: async values => {
       try {
-        await signUp(values.fullName, values.username, values.password);
-        router.push("/sign-in");
+        const data = await signIn(values.username, values.password);
+        dispatch(setAccessToken(data.accessToken));
         toast.dismiss();
-        toast.success(messages.SIGN_UP_SUCCESS_MESSAGE);
+        router.push("/chat");
       } catch (error) {
         axiosErrorHandler(error);
       }
@@ -40,24 +35,13 @@ export function SignUpForm() {
   });
 
   return (
-    <AuthForm
+    <Form
       onSubmit={formik.handleSubmit}
-      titleIcon={signUpForm.titleIcon}
-      title={signUpForm.title}
-      description={signUpForm.description}
-      buttonTitle={signUpForm.buttonTitle}
+      titleIcon={signInForm.titleIcon}
+      title={signInForm.title}
+      description={signInForm.description}
+      buttonTitle={signInForm.buttonTitle}
     >
-      <CustomInput
-        name={fullNameInput.name}
-        placeholder={fullNameInput.placeholder}
-        label={fullNameInput.label}
-        className="text-base"
-        containerClassName="w-3/5"
-        icon={fullNameInput.icon}
-        value={formik.values.fullName}
-        onChange={formik.handleChange}
-        error={formik.touched.fullName && formik.errors.fullName}
-      />
       <CustomInput
         name={usernameInput.name}
         placeholder={usernameInput.placeholder}
@@ -83,6 +67,6 @@ export function SignUpForm() {
         onBlur={formik.handleBlur}
         error={formik.touched.password && formik.errors.password}
       />
-    </AuthForm>
+    </Form>
   );
 }
