+++ frontend/src/components/Auth/auth.types.ts
@@ -1,0 +1,16 @@
+import { IconName } from "@/components/Icon/Icon";
+
+export type FormParams = {
+  title: string;
+  description: string;
+  titleIcon: IconName;
+  buttonTitle: string;
+};
+
+export type InputParams = {
+  type?: string;
+  placeholder: string;
+  name: string;
+  label: string;
+  icon: IconName;
+};
