--- frontend/src/components/AuthCard/authCard.types.ts
@@ -1,3 +1,0 @@
-import { DetailedHTMLProps, FormHTMLAttributes, PropsWithChildren } from "react";
-
-export type AuthCardProps = DetailedHTMLProps<FormHTMLAttributes<HTMLFormElement>, HTMLFormElement> & PropsWithChildren;
