+++ frontend/src/components/Card/Card.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import { AuthCardProps } from "@/components/AuthCard/authCard.types";
+import { CardProps } from "@/components/Card/card.types";
 
-export function AuthCard({ children, className, ...props }: AuthCardProps) {
+export function Card({ children, className, ...props }: CardProps) {
   return (
     <form
       className={`rounded-3xl h-fit w-[30rem] p-4 flex flex-col items-center py-20 shadow-2xl border border-gray-100 dark:bg-dark-550 dark:border-0 dark:shadow-none ${className}`}
