+++ frontend/src/components/Card/card.types.ts
@@ -1,0 +1,3 @@
+import { DetailedHTMLProps, FormHTMLAttributes, PropsWithChildren } from "react";
+
+export type CardProps = DetailedHTMLProps<FormHTMLAttributes<HTMLFormElement>, HTMLFormElement> & PropsWithChildren;
